refactor(server): move serveExportFile to its own file
